requirements.txt
fastapi==0.95.2
boto3>=1.26.0
aws-xray-sdk>=2.4.2
uvicorn[standard]==0.22.0
gunicorn
opentelemetry-sdk==1.15.0
opentelemetry-api==1.15.0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk-extension-aws==2.0.1
